text_classification/requirements.txt
pytorch
simpletransformers
pandas
torchvision